third_party/blink/web_tests/virtual/pna-navigations-disabled/external/wpt/fetch/private-network-access/window-open.tentative.https.window_include=from-treat-as-public-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public-address to local: failed preflight.
  assert_equals: expected "timeout" but got "success"
[FAIL] treat-as-public-address to local: missing CORS headers.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] treat-as-public-address to local: missing PNA header.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] treat-as-public-address to local: success.
  assert_equals: expected "success" but got "no preflight received"
[FAIL] treat-as-public-address to private: failed preflight.
  assert_equals: expected "timeout" but got "success"
[FAIL] treat-as-public-address to private: missing CORS headers.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] treat-as-public-address to private: missing PNA header.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] treat-as-public-address to private: success.
  assert_equals: expected "success" but got "no preflight received"
Harness: the test ran to completion.
